Replaced 4 AsyncTasks in LogDao
Continuing to replace the deprecated AsyncTask. This time, I found a lot to replace in one class. I found the four that were easiest to replace were the ones that merely wrote to the database, and returned nothing. Since they require no major redesign. So I quickly replaced those, before starting work on the next set.