Replaced final 2 AsyncTasks in LogDao
Finished replacing the deprecated AsyncTask. These two were a bit more complicated. Because they returned information, I had to consider the context in which the code was run. I found it easiest to just go back in the code, where each of the methods are called, until I got to a method that was returning void. Then just reorganise the code to run in a thread. I had to make sure the UI code that used the information I collected would only run afterward, but the UI code that didn't use the information would just carry on as normal. I was worried there would be issues, but they all seem to run fine, even if I pause the threaded code, everything else runs smoothly, so I can safely say there are no issues with race conditions as well.