Fixed synthetic accessor
I loosened the access level of the identifyQuestion method so it can be properly accessed by the 3 to 4 migration without the generation of a synthetic accessor. Also, removed the 'final' label because it's a static method, there's really no point to declaring it 'final'. Don't know why it was there in the first place.