Couple minor code tweaks
Made a couple minor changes to the code. Added a super call to the onActivityResult method, and removed the use of a variable in the initialize method which was only used for casting, which was really unnecessary. I also changed one of settings files so the default Nullable and NotNull annotations are the androidx ones, and not the ones from the defunct android.support library. None of this makes any changes to functionality.