Fixed whitespace issue
In some cases, the detail screen would have a lot of whitespace. This was due to the 'wrap_content' calculation being made before the 'autoSizeText' calculations. The TextSize would be shrunk, but the TextView wouldn't shrink accordingly. I fixed this by double checking the size of the box after everything's rendered. Using a bit of multithreading, I programmed a loop to keep checking if the rendering is finished, then have a second loop to check if the text is done autosizing. Once that's done running, it then resizes the height of the box to reduce the whitespace.