Rebuilt autosizing
I replaced the autosizing in TextView by creating a whole new class that is just better at autosizing. It extends the TextView class, and resizes both the font and text box size during the measuring stage, which is before it's rendered on screen, which means I was able to delete the CorrectSizeTask class, and that the ReferenceDetail dialog doesn't have to be awkwardly resized after it appears on screen. In many ways, this looks a lot better.