Replaced Twitter link with Mastodon link
Replaced the link to my Twitter account (which I no longer use) with a link to my Mastodon account. Deleted the Twitter logo and added the Mastodon logo, and reoriented the AboutScreen to fit the larger logo. Also, updated the reference to Twitter on the README to reflect the name change, as well as suggest better services since Twitter is shit now. I also updated the locale files to replace the references to Twitter with Mastodon.